requirements.txt
flask
python-Wappalyzer
validators